requirements.txt
# Google Search Console + Claude Code Integration
# Python dependencies for GSC connection testing and MCP server setup

# Google API dependencies
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# Optional: MCP framework (if building custom server)
# mcp>=0.0.1

# Optional: Async support
# httpx>=0.25.0
# anyio>=4.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# mypy>=1.0.0
